requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.3.4
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.1
django-appconf==1.0.4
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-modal-forms==2.2.0
django-bootstrap4==3.0.1
django-celery==3.3.1
django-celery-results==2.0.1
django-crispy-forms==1.11.2
django-filter==2.4.0
django-import-export==2.5.0
django-redis==4.12.1
django-select2==7.7.1
django-sendgrid-v5==0.9.1
django-tinymce==3.3.0
django-widget-tweaks==1.4.8
et-xmlfile==1.1.0
future==0.18.2
kombu==5.0.2
MarkupPy==1.14
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.7
pkg-resources==0.0.0
prompt-toolkit==3.0.18
python-http-client==3.3.2
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
sendgrid==6.7.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
tablib==3.0.0
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
